requirements.txt
fastapi
uvicorn
requests
python-dotenv
selenium
webdriver-manager
fake-useragent